Added functions to get distances to primatives and meshes
Created new file distances.jl which computes the absolute_distance and 
the signed_distance to triangle meshes, HyperSpheres and HyperRects. 
These functions have been exported.

The sphere and rect signed functions are straightforward, and there is a 
fallback for absolute_distance to equal abs(signed_distance).

Dealing with triangles and triangle meshes is less straightforward. The 
absolute distance to a triangle is a well defined (if complex) function, 
but the signed distance isn't. I've used the signed distance to the 
plane aligned with the triangle. The signed distance to a mesh of 
triangles is then the signed distance to the convex union of those 
planes. This is simple, but obviously has drawbacks.

Alternatively, the absolute distance to a mesh of triangles is robust, 
and could be supplimented with a global operation (like the winding 
number) to determine a sign. 

I've added a 27 tests to runtests which illustrate some of these issues.